DOMCore/tests/submissions/Ms2ger/Node-isEqualNode.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Node.isEqualNode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"/>
<script>
function testNullHandling(node) {
  test(function() {
    assert_false(node.isEqualNode(null))
    assert_false(node.isEqualNode(undefined))
  })
}
[
  document.createElement("foo"),
  document.createTextNode("foo"),
  document.createProcessingInstruction("foo", "bar"),
  document.createComment("foo"),
  document,
  document.implementation.createDocumentType("html", "", ""),
  document.createDocumentFragment()
].forEach(testNullHandling)
</script>
</body>
</html>
